Start!
Start of this little os project thing.

Currently just boots and shows a cat image.

Starts in kernel/boot/boot.asm
boot.asm loads the filetable, then finds the secondstage bootloader
and kernel in there and then loads those into memory.
After that it jumps to the secondstage bootloader.

secondstage.asm sets the VESA/VBE (I don't know which one i'm supposed
to use) video mode to 1280x720 with 32 bpp.
Then it loads a GDT and switches to 32-bit protected mode.
Once in 32-bit, it copies the VBE mode info to a specific address so
the kernel can easily access it and then it jumps to C kernel code.

all that the kernel currently does is just copy the cat image
(kernel/cat_image.h) to video memory so it shows up on screen!